requirements.txt
python-dotenv>=1.0.0
selenium>=4.15.2
PyMySQL>=1.1.0
requests>=2.31.0
pdfkit>=1.0.0
bs4>=0.0.1
beautifulsoup4>=4.12.2
pandas>=2.1.4
openpyxl>=3.1.2
imgkit>=1.2.3
matplotlib>=3.8.2
scipy>=1.11.4
seaborn>=0.13.0
elasticsearch>=8.11.1
urllib3>=2.1.0